% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr2sir-package.R
\docType{package}
\name{nlmixr2sir-package}
\alias{nlmixr2sir}
\alias{nlmixr2sir-package}
\title{nlmixr2sir package}
\description{
Sampling importance resampling tools for nlmixr2 fits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nlmixr2/nlmixr2sir}
  \item Report bugs at \url{https://github.com/nlmixr2/nlmixr2sir/issues}
}

}
\author{
\strong{Maintainer}: Justin Wilkins \email{justin.wilkins@occams.com} (\href{https://orcid.org/0000-0002-7099-9396}{ORCID})

Authors:
\itemize{
  \item Justin Wilkins \email{justin.wilkins@occams.com} (\href{https://orcid.org/0000-0002-7099-9396}{ORCID})
}

}
\keyword{internal}
